アクセス制限実装

+++ app/controllers/books_controller.rb
@@ -17,24 +17,30 @@ class BooksController < ApplicationController
   def index
     @books = Book.all
     @user_info = User.find(current_user.id)
-    @post = Book.new
+    @book = Book.new
   end
 
   def show
     @book = Book.find(params[:id])
-    @user_info = User.find(@book.user_id)
+    @user_info = @book.user
     @post = Book.new
   end
 
   def edit
     @book = Book.find(params[:id])
+    unless @book.user == current_user
+      redirect_to books_path
+    end
   end
 
   def update
     @book = Book.find(params[:id])
+    unless @book.user == current_user
+      redirect_to books_path
+    end
     if @book.update(book_params)
       flash[:update_book] = "You have updated book successfully."
-      redirect_to book_path(book.id)
+      redirect_to book_path(@book.id)
     else
       render :edit
     end
@@ -42,6 +48,9 @@ class BooksController < ApplicationController
 
   def destroy
     book = Book.find(params[:id])
+    unless book.user == current_user
+      redirect_to books_path
+    end
     book.destroy
     redirect_to books_path
   end
